% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annots.r
\name{annotate_gos}
\alias{annotate_gos}
\title{Annotate GO terms}
\usage{
annotate_gos(dat, colname = "term")
}
\arguments{
\item{dat}{A data.frame or tibble}

\item{colname}{Name of the column with the terms}
}
\value{
A data .frame or tibble
}
\description{
Internal function
}
