% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.r
\name{assign_allele}
\alias{assign_allele}
\title{Assign alleles bases on allele frequency and depth}
\usage{
assign_allele(
  dat,
  depth_thres = 1,
  freq_thres = 0.5,
  sequence = FALSE,
  na_rm = TRUE
)
}
\arguments{
\item{dat}{A data.frame or tibble}

\item{depth_thres}{depth_threshold}

\item{freq_thres}{freq_threshold}

\item{sequence}{Should the actual nucleodites be returned? If FALSE
the allele column will contain 'major' or 'minor'. If TRUE, the
allele column will contain the actual nucleotide}

\item{na_rm}{Should sites from samples that could not be assigned be returned?}
}
\value{
Same tibble as dat but with an allele column
}
\description{
Internal
}
