% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawas.r
\name{benchmark_imputation}
\alias{benchmark_imputation}
\title{Benchmark mice imputation}
\usage{
benchmark_imputation(geno, snp, outdir, p = 0.1, m = 5,
  verbose = FALSE, seed = NA, block_size = 100)
}
\arguments{
\item{geno}{A data table in BIMBAM format. Must contain
one row per SNP, the first three columns must be site_id,
minor_allele and major allele, followed by one column per
sample. Missing values must be encoded as NA.}

\item{snp}{A data table with snp information in BIMBAM
format. Must contain columns ID, pos and chr in that order.
Column ID must correspond to column site_id in geno.}

\item{outdir}{Output directory to write the results.}

\item{p}{Fraction of observations to hide.}

\item{m}{Number of imputations performed. See \link{mice}
documentation.}

\item{verbose}{Whether to print progress on imputation.}

\item{seed}{Seed for random sambling of data and for imputation
if needed.}

\item{block_size}{Number of markers to use to impute each marker.}
}
\value{
A list with elements r, p.imputed, res, and imputed_geno_file.
}
\description{
Hide a percent of the observations and compare
the results of imputation via \link{mice}
with the original observations.
}
