% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dos.r
\name{calculate_dos}
\alias{calculate_dos}
\title{Calculate Direction of Selection}
\usage{
calculate_dos(dat, test = TRUE, clean = TRUE)
}
\arguments{
\item{dat}{A gene by substitution type data frame or tibble.
Each row must correspond to a gene and it must have columns
Dn, Ds, Pn and Ps, corresponding to the McDonald-Kreitman
contingency table of each gene.}

\item{test}{It TRUE, a Z-test will be performed on the
DoS statistic.}

\item{clean}{If TRUE, genes where DoS is undefined (either
Dn and Ds are both zero, or Pn and Ps are both zero), will
be removed from the result.}
}
\value{
A tibble with the same columns as Dat plus a column
named DoS. If test is passed, a column named p.value will also
be included unless such a column already exists. In that case
a column with the name DoS.p.value will be included
}
\description{
Calculates the Direction of Selection (DoS) statistic from
McDonald-Kreitman contingency tables.
}
