% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fst.r
\name{calculate_fst}
\alias{calculate_fst}
\title{Calculate Fixation Index (Fst)}
\usage{
calculate_fst(
  Dat,
  map,
  method = "Weir-Cockerham",
  support_thres = 1,
  w_size = NULL,
  s_size = NULL,
  sorted = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{Dat}{A MIDAS dataset. Imported into R with \link{read_midas_data}}

\item{map}{A tibble with columns sample and Group corresponding
to sample names in MIDAS dataset and sample population respectiveley.}

\item{method}{Weir-Cockerham or Fstpool.}

\item{support_thres}{calculate_fst uses the \code{depth} slot of
the MIDAS data as a 'support' value to decide if a site in a given
sample will be considered. If it corresponds to depth, this is the
minimum coverage at a site at a given sample.}

\item{w_size}{Window size in bp. If both w_size and s_size
are numeric. Fst/Fst^{pool} values are calculated for a sliding
window.}

\item{s_size}{Window step size in bp. If both w_size and s_size
are numeric. Fst/Fst^{pool} values are calculated for a sliding
window.}

\item{sorted}{Logical indicating if sites are sorted in ascending
orderby ref_pos within each ref_id. Only used if sliding windows are calculated}

\item{verbose}{Logical indicating if progress during site level calculations
must be printed.}
}
\value{
A list with tibbles fst and w_fst corresponding to site
and window level Fst calculations. w_fst is NULL if no window
calculations are performed.
}
\description{
Calculates Fst from imported midas data
}
