% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{calculate_genome_wide_ni}
\alias{calculate_genome_wide_ni}
\title{Calculate genome-wide NI and alpha}
\usage{
calculate_genome_wide_ni(file)
}
\arguments{
\item{file}{File}
}
\value{
vector with NI and alpha
}
\description{
Reads a file from MKtest.py and uses the TG
estimator to calculate a genome-wide neutrality
index as well as alpha
}
\author{
Sur Herrera Paredes
}
