% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsites.r
\name{calculate_p_error}
\alias{calculate_p_error}
\title{Calculate prob of error}
\usage{
calculate_p_error(dat, error_prob = 0.01)
}
\arguments{
\item{dat}{A data.frame or tibble, must have columns freq
and depth}

\item{error_prob}{Baseline err probability}
}
\value{

}
\description{
Internal function used to estimate prob that minor allele
at site is the product of error only.
}
