% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{calculate_snp_dist}
\alias{calculate_snp_dist}
\title{Title}
\usage{
calculate_snp_dist(info, freq, depth, map, depth_thres = 1)
}
\arguments{
\item{info}{}

\item{freq}{}

\item{depth}{}

\item{map}{}

\item{depth_thres}{}
}
\description{
Title
}
