% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{check_map}
\alias{check_map}
\title{Process map}
\usage{
check_map(map, map_file, focal_group = NA)
}
\arguments{
\item{map}{Either a data frame or tibble or the path to
a file containing a table. Must have columns ID and Group.}

\item{map_file}{Same as map, for backwars compatibility}

\item{focal_group}{Level from Group column to use as
focal grou, i.e. compare everything else to this group.}
}
\value{
A tibble with columns sample and Group.
}
\description{
Internal
}
