% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_data.r
\name{check_midas_data}
\alias{check_midas_data}
\title{Checks loaded midas data}
\usage{
check_midas_data(Dat)
}
\arguments{
\item{Dat}{A MIDAS data}
}
\value{
TRUE if all tests pass
}
\description{
Internal
}
