% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{check_pvals_in_file}
\alias{check_pvals_in_file}
\title{Check p-values in MKtest.py outfile}
\usage{
check_pvals_in_file(file, which, plot = TRUE)
}
\arguments{
\item{which}{either 'all' or the name of the test to
be analyzed}

\item{plot}{whether to return a histogram of the p-values}
}
\value{
A named list where each element corresponds to
a test and each element is alist with pi0 and p1 entries.
See \link{check_pvalues} for more info
}
\description{
Opens a file from MKtest.py and uses the qvalue methods
of Storey and Tibshirani to estimate the proportion
of True Negatives (Pi0) in the data
}
\author{
Sur Herrera Paredes
}
