% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{check_pvalues}
\alias{check_pvalues}
\title{Check p-values}
\usage{
check_pvalues(estimates, pvals, plot = TRUE)
}
\arguments{
\item{estimates}{numeric verctor of estimates}

\item{pvals}{numeric vector of p-values}

\item{plot}{plot or not}
}
\value{
a list with elements pi0 with the point estimate
of pi0 and p1 which is a ggplot2 plot or NULL in case parameter
plot was FALSE
}
\description{
Takes a pair of vectors of estimates and their p-values
for some hypothesis test, and uses qvalue methods from
storey and Tibshirani to estimate the number of True
Negatives (pi0) and creat a p-value histogram. It doesn't
consider cases where the estimate is undefined.
}
\author{
Sur Herrera Paredes
}
