% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsites.r
\name{determine_sample_dist}
\alias{determine_sample_dist}
\title{Determine distributiion within sample}
\usage{
determine_sample_dist(dat, thres = 0.05, error_prob = 0.01)
}
\arguments{
\item{dat}{A data.frame or tibble. One site per line,
must have columns depth and freq.}

\item{thres}{p-value threshold to consider that the evidence
against error only model is too large. Sites within samples
with p-value below threshold will be considered heterogeneous.}

\item{error_prob}{Baseline error probability in sequences.}
}
\value{
A data frame or tibble with columns for depth, freq
and sample_dist, as well as any other columns previously present.
}
\description{
For each site within each sample, it determines if
the dominant allele is homogeneous (likeley the only
allele present), or if there is evidence of a
second allele present (heterogeneous).
}
