% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsites.r
\name{determine_substitution_type}
\alias{determine_substitution_type}
\title{Determine DNA substitution type}
\usage{
determine_substitution_type(info, clean = TRUE)
}
\arguments{
\item{info}{A data.frame or tibble. It should have columns
'major_allele' and 'minor_allele'. These columns must have
single DNA characters in upper case.}

\item{clean}{Indicate if sites should be removed when
substitution type could not be correctly inferred.}
}
\value{
The same data.frame or tibble as info, with an extra
column called 'substitution'.
}
\description{
Determines if a substitution was made of transitions or
transversions.
}
\examples{
library(HMVAR)

i <- dplyr::tibble(major_allele = c('A', 'C', 'G', 'T', 'a', NA),
                   minor_allele = c('T', 'T', 'A', 'A', 'T', 'G'))
determine_substitution_type(i, clean = FALSE)
}
