% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dos.r
\name{dos}
\alias{dos}
\title{Calculate Direction of Selection from SNV data}
\usage{
dos(
  info,
  freq,
  depth,
  map,
  depth_thres = 1,
  freq_thres = 0.5,
  test = TRUE,
  clean = TRUE
)
}
\arguments{
\item{info}{Data table corresponding to the 'snps_info.txt'
file from MIDAS. Must have columns 'site_id' and 'sample'}

\item{freq}{A data table corresponding to the 'snps_freq.txt'
file from MIDAS. Must have a 'site_id' column, and one more
column per sample. Each row is the frequency of the minor
allele for the corresponding site in the corresponding sample.}

\item{depth}{A data table corresponding to the 'snps_depth.txt'
file from MIDAS. Must have a 'site_id' column, and one more
column per sample. Each row is the sequencing depth for the
corresponding site in the corresponding sample.}

\item{map}{A data table associating samples with groups (sites).
must have columns 'sample' and 'Group'.}

\item{depth_thres}{Minimum number of reads (depth) at a site at
a sample to be considered.}

\item{freq_thres}{Frequency cuttoff for minor vs major allele.
The value represents the distance from 0 or 1, for a site to be
assigned to the major or minor allele respectively. It must be
a value in [0,1].}

\item{test}{It TRUE, a Z-test will be performed on the
DoS statistic.}

\item{clean}{If TRUE, genes where DoS is undefined (either
Dn and Ds are both zero, or Pn and Ps are both zero), will
be removed from the result.}
}
\value{
A tibble with the same columns as Dat plus a column
named DoS. If test is passed, a column named p.value will also
be included unless such a column already exists. In that case
a column with the name DoS.p.value will be included
}
\description{
Takes SNV data in MIDAS format and produces Direction
of Selection (DoS) statistics.
}
