% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annots.r
\name{expand_annot}
\alias{expand_annot}
\title{Expand annotation}
\usage{
expand_annot(gene_id, terms)
}
\arguments{
\item{gene_id}{Character}

\item{terms}{Character string of comma-separated terms for the
`gene_id`.}
}
\value{
A tibble with one row per term
}
\description{
Internal function
}
