% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singletons.r
\name{find_singletons}
\alias{find_singletons}
\title{Find singletons}
\usage{
find_singletons(alleles, midas_dir, info, meta)
}
\arguments{
\item{alleles}{site x sample alleles table. Must have column site_id plus
one column per sample}

\item{midas_dir}{Instead of alleles table. A path with output files
in the format of midas merge can be provided. Must have freqs, depth and
info files.}

\item{info}{Optional site info table to append metadata to singleton
sites. Must have a site_id column.}

\item{meta}{Optional sample metadata to append to samplew with singleton
sites. Must have column sample.}
}
\value{
a tibble with one row per singleton
}
\description{
Finds variants that have singleton alleles and returns the
positions and the samples where they ocurr.
}
