% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{find_snp_effect}
\alias{find_snp_effect}
\title{Determine the effect of a coding variant on the aminoacid sequence}
\usage{
find_snp_effect(info, nucleotides = c(A = 1, C = 2, G = 3, T = 4))
}
\arguments{
\item{info}{A data table corresponding to the contents of
the snp_info.txt file produced by midas_merge. It must have columns:
'site_id', 'major_allele', 'minor_allele' and 'amino_acids'. The aminoacid
column must contain a string of four comma-separated values indicating
the aminoacid encoded for each variant (eg. 'V,L,V,L').}

\item{nucleotides}{Named vector indicating the position of each
nucleotide that corresponds to the amino_acids column in info. The
default corresponds to MIDAS v1.3.1 default.}
}
\value{
The same data table passed as info with a factor column
'snp_effect' added.
}
\description{
Takes an table corresponding to the contents of the snp_info.txt
file from midas_merge.py and adds a column indicating whether
the variant is synonymous or non-synonymous
}
