% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIT.r
\name{FIT}
\alias{FIT}
\title{Frequency Increment Test (FIT)}
\usage{
FIT(Dat, aggregate = TRUE, ...)
}
\arguments{
\item{Dat}{One of three options:
1. A list from \link{read_midas_data}
2. A tibble corresponding to the freq element of 1.
3. A tibble with all the in long format must have site_id, freq, pop_id
& time columns}

\item{aggregate}{}

\item{map}{For 1 and 2 a map indicating sample population and timepoint}

\item{min_depth}{For 1, min depth to consider a site.}
}
\value{
A tibble. Empty if no site x pop_id combination has at least three
timepoints
}
\description{
From https://www.genetics.org/content/196/2/509
}
