% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.r
\name{format_input}
\alias{format_input}
\title{Formats input form HMMCP 16S files}
\usage{
format_input(name, counts_file, taxonomy_file, Map = Map,
  collapse_level = NULL)
}
\value{
A Dataset object
}
\description{
Only to be used to process 16S files from MOTHUR \
pipeline of HMP (HMMCP). Count files must be pre-edited
to remove trailing tab
}
\author{
Sur Herrera Paredes
}
