% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.r
\name{format_input_hmqcp}
\alias{format_input_hmqcp}
\title{Read HMQCP input}
\usage{
format_input_hmqcp(name, counts_file, map_file, collapse_level = 6)
}
\arguments{
\item{name}{name}

\item{counts_file}{counts file}

\item{map_file}{map file}

\item{collapse_level}{collapse level}
}
\description{
Reads QIIME outputs from HMP files of HMQCP dataset.
and produces a genus level dataset for the fgour body sites
of interest.
}
\author{
Sur Herrera Paredes
}
