% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawas.r
\name{gemma_kinship}
\alias{gemma_kinship}
\title{Calculate kinship matrix with GEMMA}
\usage{
gemma_kinship(geno_file, pheno_file, snp_file, gemma = "gemma",
  outdir = "kinship/", prefix = "kinship")
}
\arguments{
\item{geno_file}{Mean genotype file in BIMBAM format. See GEMMA
manual for details.}

\item{pheno_file}{Phenotype file in BIMBAM format. See GEMMA
manual for details.}

\item{snp_file}{SNP annotation file in BIMBAM format. See GEMMA
manual for details.}

\item{gemma}{GEMMA executable.}

\item{outdir}{directory to write output.}

\item{prefix}{Prefix for output filenames.}
}
\description{
Calculate kinship matrix with GEMMA
}
\references{
http://www.xzlab.org/software/GEMMAmanual.pdf
}
