% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawas.r
\name{gemma_lmm}
\alias{gemma_lmm}
\title{Run LMM via GEMMA}
\usage{
gemma_lmm(
  geno_file,
  pheno_file,
  snp_file,
  kinship_file,
  cov_file = NULL,
  gemma = "gemma",
  outdir = "lmm/",
  maf = 0,
  prefix = "lmm"
)
}
\arguments{
\item{geno_file}{Mean genotype file in BIMBAM format. See GEMMA
manual for details.}

\item{pheno_file}{Phenotype file in BIMBAM format. See GEMMA
manual for details.}

\item{snp_file}{SNP annotation file in BIMBAM format. See GEMMA
manual for details.}

\item{kinship_file}{Kinship file in GEMMA format. See GEMMA
manual for details.}

\item{cov_file}{Covariates file in BIMBAM format. See GEMMA
manual for details.}

\item{gemma}{GEMMA executable.}

\item{outdir}{directory to write output.}

\item{maf}{Minor allele frequency threshold for SNPs to test.}

\item{prefix}{Prefix for output filenames.}
}
\description{
Tested with GEMMA 0.93b modified by bugwas package.
}
\references{
http://www.xzlab.org/software/GEMMAmanual.pdf
}
