% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_data.r
\name{gene_midas_data}
\alias{gene_midas_data}
\title{Get all MIDAS data from a gene}
\usage{
gene_midas_data(Dat, gene, depth_thres = 1, freq_thres = 0.5)
}
\arguments{
\item{Dat}{A list. See output from \link{read_midas_data}}

\item{gene}{Name of the gene to be extracted, must exist in
'gene_id' column of the 'info' element from Dat.}

\item{depth_thres}{Minimum depth trheshold.}

\item{freq_thres}{Freq threshold.}
}
\value{
A tibble that results of merging the info, dat and freq elements
of Dat. Plus allele assignments.
}
\description{
Returns a list similar to the one by \link{read_midas_data} but
only with sites from a given gene.
}
