% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.r
\name{gene_snv_aln}
\alias{gene_snv_aln}
\title{Get gene SNV alignment}
\usage{
gene_snv_aln(snps, seq, missing_as = "major", strand = "+",
  keep_last_codon = TRUE)
}
\arguments{
\item{snps}{A tibble or data.frame. It must have columns
'site_id', 'sample', 'ref_pos', 'ref_allele', 'major_allele',
and 'allele'. Column 'minor_allele' is not required at the moment
but might be in the future.}

\item{seq}{A character vector with the sequence of the fragment.}

\item{missing_as}{How to treat sites in samples when there is no
information in snps? Either 'N'(set all those sites to N), 'major'
(use the major allele for that site), or 'ref' (use the reference allele).}

\item{strand}{If strand is '-', then the sequence will be reverse complemented.}

\item{keep_last_codon}{This is useful to remove stop codons when gene is a
CDS. It will trim the last 3 sites from the alignment.}
}
\value{
An algnment object from \link{seqinr}
}
\description{
Takes SNP informmation from multiple samples for a sequence
fragment, and a reference sequence for that fragment. It produces
and 'alignment' object that has one 'aligned' sequnece per sample
where the SNPs have been replaced.
}
