% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide_ni.r
\name{genome_wide_ni}
\alias{genome_wide_ni}
\title{Calculate genome-wide NI and alpha}
\usage{
genome_wide_ni(file, col_types = "ccnnnnnnnnn", na = "nan")
}
\arguments{
\item{file}{File to a tab-delimited file. Should have Dn,
Ds, Pn, and Ps columns, and one row per gene.}

\item{col_types}{Column definitions for \link[readr]{read_tsv}.}

\item{na}{Vector corresponding to NaN values.}
}
\value{
vector with NI and alpha
}
\description{
Reads a file from MKtest.py and uses the TG
estimator to calculate a genome-wide neutrality
index as well as alpha
}
\details{
Uses estimatro NI_TG as fefined by Stoletzki & Eyre-Walker 2011.
Based on work from Tarone 1981 and Greenland 1982.

\deqn{NI_{TG} = \frac{\sum D_{si}P_{ni}/(P_{si} + D_{si})}{\sum P_{si}D_{ni}/(P_{si} + D_{si})}}

where i is the i-th gene.
}
\author{
Sur Herrera Paredes
}
