% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{get_mk_results_files}
\alias{get_mk_results_files}
\title{Get files for MK results}
\usage{
get_mk_results_files(d, pattern = "^mk_results")
}
\arguments{
\item{d}{Directory}

\item{pattern}{Pattern to match for files
containing results from MKtests.py}
}
\value{
Vector of files
}
\description{
Takes a directory and finds all files that
correspond to results of MKtests.py, and returns
a list of full paths for those files
}
\author{
Sur Herrera Paredes
}
