% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{get_site_dist}
\alias{get_site_dist}
\title{Determine whether a variant site is fixed or polymorphic}
\usage{
get_site_dist(d, group_thres = 2)
}
\arguments{
\item{d}{A data table that matches each site at each
sample with the group of that sample. Must have column
'allele' which must have two values only; column 'Group'
which is the grouping factor (can be character) with two
levels.}

\item{group_thres}{The minimum number of samples per
group to consider the site.}
}
\value{
A character string indicating the type of site:
'Fixed', 'Polymorphic', 'Invariant' or NA.
}
\description{
This is an internal utility function
}
\details{
Looks at the distribution of alleles at one site and
determines how it distributes with respect to a grouping factor.
If the site perfectly separates both groups it is 'Fixed'
and 'Polymorphic' otherwise.

Assumes grouping factor with two levels (NO CHECKS!!)

Assumes bi-allelic site indicated by character string (NO CHECKS!!)

Assumes only data from one site is give (NO CHECKS!!)

USE WITH CARE!!
}
