% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsites.r
\name{group_size}
\alias{group_size}
\title{Size of groups}
\usage{
group_size(d, columns)
}
\arguments{
\item{d}{A data frame or tibble}

\item{columns}{Names of columns to group}
}
\value{
A tibble
}
\description{
Internal function
}
