% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annots.r
\name{gsea}
\alias{gsea}
\title{Title}
\usage{
gsea(dat, test = "wilcoxon", alternative = "greater", min_size = 3)
}
\arguments{
\item{dat}{}

\item{test}{}

\item{alternative}{}

\item{min_size}{}
}
\description{
Title
}
