% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ld_r.r
\name{ld_r}
\alias{ld_r}
\title{Calculate linkage disequilibrium (LD) R^2}
\usage{
ld_r(x)
}
\arguments{
\item{x}{A SNP numerical matrix with SNPs as rows & samples
as columns}
}
\value{
A distance object
}
\description{
Calculates r^2 = D ^2 / (pA(1-PA)pB(1-pB))
}
