% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawas.r
\name{mice_impute}
\alias{mice_impute}
\title{Imputation via mice}
\usage{
mice_impute(geno, snp, outdir = "imputed/", m = 5, verbose = FALSE,
  prefix = "imputed", return_table = FALSE, seed = NA,
  block_size = 100)
}
\arguments{
\item{geno}{A data table with genotype information. First three columns
must be site_id, minor_allele and major allele; folowed by one column per
sample with the sample ID as column name. Missing genotypes must be encoded
as NA values}

\item{snp}{SNP information data table. Must have columns ID, pos and chr.
ID mut correspond to the "site_id" column in \code{geno}, and the SNPs must
be in the same order in both tables.}

\item{outdir}{Output directory to create and write the imputed table in
BIMBAM format.}

\item{m}{Number of multiple imputations. See \link{mice} documentation.}

\item{verbose}{Whether to print progress during imputation.}

\item{prefix}{Prefix for imputed file.}

\item{return_table}{If TRUE, the function will return a list with both
the output file path and the imputed table. If FALSE, only the file path
will be returned.}

\item{seed}{Seed for imputation. See \link{mice} documentation.}

\item{block_size}{Number of markers to use to impute each marker.}
}
\value{
Either the output file path, or a list containing the file path
and the imputed table.
}
\description{
Imputes missing genotypes using mice
}
