% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dos.r
\name{midas_dos}
\alias{midas_dos}
\title{Direction of Selection statistics on MIDAS output}
\usage{
midas_dos(
  midas_dir,
  map,
  map_file,
  genes = NULL,
  depth_thres = 1,
  freq_thres = 0.5,
  focal_group = NA
)
}
\arguments{
\item{midas_dir}{Directory where the output from
midas_merge.py is located. It must include files:
'snps_info.txt', 'snps_depth.txt' and 'snps_freq.txt'.}

\item{map}{Either  a file path or a tibble. If a path,
it mos point to a mapping file associating samples
in the MIDAS otput to groups. It must have an 'ID'
and a 'Group' column. If a tibble. It must have columns
'sample' and 'Group'.}

\item{map_file}{Same as map. Present for backwards compatibility.}

\item{genes}{The list of genes that are to be tested.
Must correspond to entries in the 'genes_id' column
of the 'snps_info.txt' file. If NULL, all genes will
be tested.}

\item{depth_thres}{The minimum number of reads at
a position in a given sample for that position in that
sample to be included.}

\item{freq_thres}{Frequency cuttoff for minor vs major allele.
The value represents the distance from 0 or 1, for a site to be
assigned to the major or minor allele respectively. It must be
a value in [0,1].}

\item{focal_group}{A string indicating the group that should
be compared to everything else. If different from NA, all values
in the 'Group' column of the mapping file will be converted to
"non.<focal_group>", ensuring a dichotomous grouping.}
}
\value{
A tibble with columns gene_id, Dn, Ds, Pn, Ps, DoS and p.value.
}
\description{
Take output from MIDAS script midas_merge.py
and obtain and test Direction of Selection (DoS)
statistics.
}
\examples{
midas_dos(midas_dir = system.file("toy_example/merged.snps/", package = "HMVAR"),
          map = system.file("toy_example/map.txt", package = "HMVAR"))
}
