% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{mktest}
\alias{mktest}
\title{McDonald-Kreitman test}
\usage{
mktest(alleles, info, map)
}
\arguments{
\item{alleles}{site x sample alleles table. First column must be
"site_id"}

\item{info}{Table with information about SNVs. Must include columns
"site_id", "locus_type", "gene_id" and "snp_effect".}

\item{map}{Mapping file. must have columns "sample" and "group"}
}
\value{
A tibble
}
\description{
McDonald-Kreitman test
}
