% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{mkvalues}
\alias{mkvalues}
\title{Entries in McDonald-Kreitman table.}
\usage{
mkvalues(info)
}
\arguments{
\item{info}{A data table corresponding to the
MIDAS snps_info.txt output file. Must have been
processed with \link{determine_snp_effect} and
\link{determine_snp_dist}, and so it must contain
columns 'snp_effect' and 'distribution'}
}
\value{
A tibble with one column with the count
of each substitution type.
}
\description{
Calculates the four entries (Dn, Ds, Pn, Ps)
in the McDonald-Kreitman contingency table.
}
