% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotype2rdp.r
\name{phylotype2rdp}
\alias{phylotype2rdp}
\title{Format HMMCP phylotype taxonomy for AMOR}
\usage{
phylotype2rdp(x, split.char = ";")
}
\arguments{
\item{x}{A character vector, one element per taxa}

\item{split.char}{A character string indicating the field
delimiter character}
}
\value{
A character vector
}
\description{
Format HMMCP phylotype taxonomy for AMOR
}
\author{
Sur Herrera Paredes
}
