% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylotype2rdp.r
\name{phylotype2rdp2}
\alias{phylotype2rdp2}
\title{Phylotype to RDP version 2}
\usage{
phylotype2rdp2(x, split.char = ";")
}
\arguments{
\item{x}{A character vector, one element per taxa}

\item{split.char}{A character string indicating the field
delimiter character}
}
\description{
Slow version
}
