% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_qqplot.r
\name{pval_qqplot}
\alias{pval_qqplot}
\title{p-value qqplot}
\usage{
pval_qqplot(pvector, main = NULL, ...)
}
\arguments{
\item{pvector}{Vector of p-values}

\item{main}{Title}

\item{...}{Extra parameters for plot}
}
\value{
A base plot
}
\description{
quantile-quantile plot of expected vs observed
p-values. With base graphics.
}
\details{
Converts zeroes to the smalles non-zero p-value which
is required for the log-transformation. It uses base
graphics becasue ggplot2 can be very inefficient when
there are millions of p-values.
}
\references{
\url{http://GettingGeneticsDone.blogspot.com/}
\url{http://gettinggeneticsdone.blogspot.com/p/copyright.html}
}
