% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_genotypes.r
\name{qp_genotypes}
\alias{qp_genotypes}
\title{Quasi-phased genotypes}
\usage{
qp_genotypes(
  midas_dir,
  midas_dat,
  map = NULL,
  min_depth = 5,
  min_snv_prop = 0.8,
  maf_thres = 0.2
)
}
\arguments{
\item{midas_dir}{Directory with outptu from mdas_merge.py snps. Either
this or midas_dat needs to be provided.}

\item{midas_dat}{A list with info, freq, and depth elements corresponding
to the result of read_midas_data. Either this or midas_dir needs to be
provided}

\item{map}{A data table associating samples in the MIDAS otput to groups.
It must have an 'sample' and a 'Group' column. Only used if midas_dat was
passed and only samples in map are processed.}

\item{min_depth}{Minimum sequence depth at a given sample and position
to be abele to call a gnotype}

\item{min_snv_prop}{Minimum proportion of SNVs in a sample, among SNVs that
pass the min_depth filter, that pass the maf_thres filter. Samples that
have a smaller proportion of SNVs passing this condition are considered
non-quasi-phaseable and genotype calls are not made.}

\item{maf_thres}{A value in the range [0, 0.5]. SNVs are only considered 
quasi-phaseable if their frequency (f) fullfills f <= maf_thres | f >= 1 -
maf_thres, and if they are in a quasi-phaseable sample. SNVs in
quasi-phaseable samples, that do not pass the maf_thres or min_depth
filters are considered missing data.}
}
\value{
A tiblle with columns site_id, and one column per quasi-phaseable
sample, which contain the genotype calls.
}
\description{
From allele frequency data in the format of midas_merge.py snps it determines
which samples and which positions in each sample are quasi-phaseable and
makes the corresponding genotype calls.
}
