% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_eggnog.r
\name{read_eggnog}
\alias{read_eggnog}
\title{Read eggNOG mapper annotations}
\usage{
read_eggnog(path)
}
\arguments{
\item{path}{File path}
}
\value{
A tibble
}
\description{
Read eggNOG mapper annotations
}
