% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_data.r
\name{read_midas_abun}
\alias{read_midas_abun}
\title{Read MIDAS abundance file}
\usage{
read_midas_abun(file)
}
\arguments{
\item{file}{File path}
}
\value{
A tibble
}
\description{
Reads either the snps_depth.txt or snps_freq.txt
file produced by midas_merge.py
}
