% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_data.r
\name{read_midas_data}
\alias{read_midas_data}
\title{Read midas snp merge data}
\usage{
read_midas_data(midas_dir, map, genes = NULL, cds_only = TRUE)
}
\arguments{
\item{midas_dir}{Directory where the output from
midas_merge.py is located. It must include files:
'snps_info.txt', 'snps_depth.txt' and 'snps_freq.txt'.}

\item{map}{A data table associating samples
in the MIDAS otput to groups. It must have an 'sample'
and a 'Group' column.}

\item{genes}{The list of genes that are to be tested.
Must correspond to entries in the 'genes_id' column
of the 'snps_info.txt' file. If NULL, all genes will
be tested.}

\item{cds_only}{If TRUE, it will remove all sites that
do not have a gene_id. If a gene list is passed its effect
is redundant.}
}
\value{
A list with three data tables: info, freq
and depth, corresponding to the three files from 
MIDAS
}
\description{
Reads the output of midas_merge.py
}
\examples{
library(HMVAR)
library(tidyverse)

# Get file paths
midas_dir <- system.file("toy_example/merged.snps/", package = "HMVAR")
map <- read_tsv(system.file("toy_example/map.txt", package = "HMVAR"),
                col_types = cols(.default = col_character())) \%>\%
  select(sample = ID, Group)

# Read data
midas_data <- read_midas_data(midas_dir = midas_dir, map = map, cds_only = TRUE)
midas_data
}
