% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_data.r
\name{read_midas_info}
\alias{read_midas_info}
\title{Read MIDAS snps_info.txt file}
\usage{
read_midas_info(file)
}
\arguments{
\item{file}{Path to file}
}
\value{
A tibble with the contents of the 
snps_info.txt file from MIDAS
}
\description{
Read MIDAS snps_info.txt file
}
