% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_coef.r
\name{s_coefficient}
\alias{s_coefficient}
\title{Calculate selection coefficient}
\usage{
s_coefficient(Dat, ...)
}
\arguments{
\item{Dat}{There are 3 options:

1. A list like the one produced by \link{HMVAR::read_midas_data}. It
must have freq and depth elements.
2. A data frame or tibble corresponding to the freq element of option 1.
3. A data frame or tibble already including the initila (v0) and final (v1)
frequencies of every allele. It must have columns "site_id", "v0",and "v1".
For option 1 and 2, and additonal map is required with time information
for the samples. For option 3 an additional time parameter is needed.}

\item{aggregate}{Whether to calculate aggregated statistics from
multiple populations}

\item{time}{Only used (and required) for option 3 in Dat. Either a single
numeric value indicating the time between the first and last allele
frequency measurement, or a single character value indicating the name of
the column containing the time between first and last allele frequency
measurement.}

\item{map}{}
}
\value{
A tibble with selection coefficient calculations
}
\description{
Selection coefficient for alleles that change frequency between two
time points.
}
\details{
According to https://www.genetics.org/content/196/2/509
}
