% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fst.r
\name{site_fst}
\alias{site_fst}
\title{Fixation Index (Fst) for one site}
\usage{
site_fst(
  freq,
  support,
  info,
  map,
  support_thres = 1,
  method = "Weir-Cockerham"
)
}
\arguments{
\item{freq}{A 1 x (n+1) tibble. One of the columns must correspond
to 'site_id', and the rest of the columns correspond to one sample
each. Entries under each sample correspond to allele frequencies.}

\item{support}{A 1 x (n+1) tibble. Columns must match columns in freq.
Values must be numeric and indicate support of each allele frequencuy}

\item{info}{A tiblle with one row. One column must correspond to
'site_id'. The rest of the columns contain information about the site}

\item{map}{A tibble with columns 'sample' (corresponding to column names)
in freq and support, and 'Group' corresponding to the populations used
for Fst calculations.}

\item{support_thres}{Minimum value in support tibble to keep a sample
for this site}

\item{method}{The method to calculate Fst. Currently only the
'Weir-Cockerham' method from 1984, and the 'Fstpool' methods are implemented.}
}
\value{
A tibble
}
\description{
Takes allele frequency data from one site only
and calculates Wright's fixation index (Fst).
}
\details{
It works only for haploid individuals (e.g proportion
of heterozygous is 0).
}
