% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annots.r
\name{test_go}
\alias{test_go}
\alias{test_go.character}
\alias{test_go.numeric}
\title{Gene Ontology enrichment via topGO}
\usage{
test_go(genes, annots, ontology, description, algorithm, statistic,
  node_size, score_threshold)

\method{test_go}{character}(genes, annots, ontology = "BP",
  description = "", algorithm = "classic", statistic = "fisher",
  node_size = 3)

\method{test_go}{numeric}(genes, annots, ontology = "BP",
  description = "", algorithm = "classic", statistic = "fisher",
  node_size = 3, score_threshold = 0.05)
}
\arguments{
\item{genes}{Either a character vector with the gene
identifiers that are 'significant' or a named numeric vector
where the vector names are the gene_identifiers of the universe
of genes and the numeric values the genes' scores.}

\item{annots}{Either a data.frame or tibble that has 'gene_id' and
'terms' column or the result of running \link{annots_to_geneGO} on
such table.}

\item{ontology}{Which ontology to test. See help at \link[topGO]{topGOdata-class}.}

\item{description}{Description for the test. See help at \link[topGO]{topGOdata-class}.}

\item{algorithm}{Algortithm for test. See help at \link[topGO]{runTest}.}

\item{statistic}{Statistic to test. See help at \link[topGO]{runTest}.}

\item{node_size}{Minimum number of genes per term to test that
term. See help at \link[topGO]{topGOdata-class}}

\item{score_threshold}{If genes is a numeric vector, then
this should be the 'significance' threshold. E.g. if scores are p-values
a common threshold would be 0.05.}
}
\value{
A list with elements topgo_data and topgo_res of class
topGOdata and topGOresult respecitveley
}
\description{
Performs Gene Ontology (GO) enrichment analysis via
topGO
}
