% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/singletons.r
\name{test_singleton_enrichment}
\alias{test_singleton_enrichment}
\title{Test for singleton enrichments}
\usage{
test_singleton_enrichment(Dat, info)
}
\arguments{
\item{Dat}{A table with one singelton site per row. Must have columns
site_id, gene_id and snp_effect.}

\item{info}{A table with information about all SNV sites (singleton and
non-singleton) in the genes
being tested. Must have columns site_id, gene_id, and snp_effect.}
}
\value{
A tibble
}
\description{
Takes table of singletons and overall SNVs and tests for an enrichment
of overall singletons of enrichments per gene, as well as an erichment
of non-synonymous over synonymous mutations among singletons versus
among non-singletons
}
