% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varsites.r
\name{determine_substitution_type}
\alias{determine_substitution_type}
\title{Determine DNA substitution type}
\usage{
determine_substitution_type(info, clean = TRUE)
}
\arguments{
\item{info}{A data.frame or tibble. It should have columns
'major_allele' and 'minor_allele'. These columns must have
single DNA characters in upper case.}

\item{clean}{Indicate if sites should be removed when
substitution type could not be correctly inferred.}
}
\value{
The same data.frame or tibble as info, with an extra
column called 'substitution'.
}
\description{
Determines if a substitution was made of transitions or
transversions.
}
\examples{
library(HMVAR)

# Get file paths
midas_dir <- system.file("toy_example/merged.snps/", package = "HMVAR")
map <- readr::read_tsv(system.file("toy_example/map.txt", package = "HMVAR"),
                       col_types = readr::cols(.default = readr::col_character())) \%>\%
  dplyr::select(sample = ID, Group)

# Read data
midas_data <- read_midas_data(midas_dir = midas_dir, map = map, cds_only = TRUE)

determine_substitution_type(midas_data$info, clean = FALSE)
}
