% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{midas_mktest}
\alias{midas_mktest}
\title{Title}
\usage{
midas_mktest(midas_dir, map_file, genes, depth_thres = 1)
}
\arguments{
\item{midas_dir}{}

\item{map_file}{}

\item{genes}{}

\item{depth_thres}{}
}
\description{
Title
}
