% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/midas_data.r
\name{midas_to_bimbam}
\alias{midas_to_bimbam}
\title{Convert MIDAS merge output to BIMBAM input}
\usage{
midas_to_bimbam(midas_dir, map, outdir, prefix = NULL)
}
\arguments{
\item{midas_dir}{Directory where midas merge output for one genome
is located. Must contain files snps_info.txt, snps_depth.txt and
snps_freq.txt}

\item{map}{Data frame or tibble that maps samples to groups. It
must have columns 'sample' and 'ID'.}

\item{outdir}{Directory where to store the results. It will be
created if it does not exists already. If it exists, and files
with the output file names exist, they will be overwriten.}

\item{prefix}{Prefix to append to all filenames.}
}
\value{
A list with elements filenames and Dat. The first element
contains the relative paths to the three BIMBAM files, and the
second contains tibbles with the data written to those files
}
\description{
Convert MIDAS merge output to BIMBAM input
}
