% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mktest.r
\name{select_samples_from_abun}
\alias{select_samples_from_abun}
\title{Select samples that are present in mapping file}
\usage{
select_samples_from_abun(abun, map)
}
\arguments{
\item{abun}{A data table where the first column is called 'site_id', and all
the other columns correspond to sample names}

\item{map}{A data table where there is a column called 'sample' which
corresponds to the column names of 'abun'.}
}
\value{
A data table
}
\description{
Select samples that are present in mapping file
}
