% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annots.r
\name{term_gsea}
\alias{term_gsea}
\title{Title}
\usage{
term_gsea(genes, scores, test = "wilcoxon", alternative = "greater",
  min_size = 3)
}
\arguments{
\item{genes}{}

\item{scores}{}

\item{test}{}

\item{alternative}{}

\item{min_size}{}
}
\description{
Title
}
