% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annots.r
\name{terms_enrichment}
\alias{terms_enrichment}
\title{Enrichment analysis of annotation terms}
\usage{
terms_enrichment(dat, method = "gsea", ...)
}
\arguments{
\item{dat}{A data.frame or tibble. It must have columns 'gene_id',
'terms' and 'score'. Values in 'gene_id' must be unique.}

\item{method}{Either 'gsea' or 'test_go'.}

\item{...}{Extra parameters for \link{gsea} or \link{test_go}. If
`method == 'test_go'` parameters 'genes', 'scores' and 'ontology'
are already provided by this function.}
}
\value{
A tibble with columnns 'term', 'size', 'statistic', and
'p.value'. If all the terms have GO-like IDs, then columns 'ontology'
and 'annotation' will be otbained from GO.db.
}
\description{
Takes a data.frame or tibble with one gene per row, annotations
per gene, and a score per gene. It performs enrichment analysis
via \link{gsea} or \link{test_go}.
}
