% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metawas.r
\name{tidy_mice}
\alias{tidy_mice}
\title{Tidy mice}
\usage{
tidy_mice(d, m = 5, verbose = FALSE, seed = NA, block_size = 100)
}
\arguments{
\item{d}{}

\item{m}{}

\item{verbose}{}
}
\value{
A tibble with imputed results
}
\description{
Internal utility function
}
\details{
Calls mice on data table
}
