% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fst.r
\name{window_fst}
\alias{window_fst}
\title{Sliding window Fst calculation}
\usage{
window_fst(dat, w_size = 1000, s_size = 300, sorted = FALSE)
}
\arguments{
\item{dat}{A tibble with columns 'ref_id', 
'ref_pos', 'a', 'b' and 'c'.}

\item{w_size}{Window size in bp.}

\item{s_size}{Step size in bp.}

\item{sorted}{logical indicating if the sites in dat.
are already in sorted ascending order by ref_pos.}
}
\value{
A tibble
}
\description{
Takes a tibble that has the a, b, and c parameters of
for Fst calculation (Weir & Cockerham 1984) and calculates
Fst over a sliding window. It only worts for a single
contig.
}
